Refactor Garbage Code

+++ WindSpigot-Server/src/main/java/net/minecraft/server/MobEffectAttackDamage.java
@@ -10,7 +10,7 @@ public class MobEffectAttackDamage extends MobEffectList {
 	public double a(int i, AttributeModifier attributemodifier) {
 		// PaperSpigot - Configurable modifiers for strength and weakness effects
 		return this.id == MobEffectList.WEAKNESS.id
-				? (double) (org.github.paperspigot.PaperSpigotConfig.weaknessEffectModifier * (i + 1))
+				? (org.github.paperspigot.PaperSpigotConfig.weaknessEffectModifier * (i + 1))
 				: org.github.paperspigot.PaperSpigotConfig.strengthEffectModifier * (i + 1);
 	}
 }
